+++ Typewriter.svelte
@@ -12,21 +12,23 @@
 
 	const sleep = ms => new Promise(resolve => setTimeout(resolve, ms))
 	const rng = (min, max) => Math.floor(Math.random() * (max - min) + min)
-	const hasSingleTextNode = el =>
-		el.childNodes.length === 1 && el.childNodes[0].nodeType === 3
-	const typingInterval = async () =>
-		await sleep(interval[rng(0, interval.length)] || interval)
+	const hasSingleTextNode = el => el.childNodes.length === 1 && el.childNodes[0].nodeType === 3
+	const typingInterval = async () => await sleep(interval[rng(0, interval.length)] || interval)
 
 	const getElements = parentElement => {
-		const treeWalker = document.createTreeWalker(
-			parentElement,
-			NodeFilter.SHOW_ELEMENT
-		)
+		const treeWalker = document.createTreeWalker(parentElement, NodeFilter.SHOW_ELEMENT)
 		let currentNode = treeWalker.nextNode()
 		while (currentNode) {
 			const text = currentNode.textContent.split('')
 			hasSingleTextNode(currentNode) && elements.push({ currentNode, text })
 			currentNode = treeWalker.nextNode()
+		}
+		if (hasSingleTextNode(node)) {
+			const text = node.textContent.split('')
+			node.textContent = ''
+			const childNode = document.createElement('p')
+			node.appendChild(childNode)
+			elements.push({ currentNode: childNode, text })
 		}
 	}
 
@@ -35,8 +37,7 @@
 		currentNode.classList.add('typing')
 		for (const letter of text) {
 			currentNode.textContent += letter
-			const fullyWritten =
-				mode === 'loop' && currentNode.textContent === text.join('')
+			const fullyWritten = mode === 'loop' && currentNode.textContent === text.join('')
 			if (fullyWritten) {
 				await sleep(loopInterval)
 				while (currentNode.textContent !== '') {
@@ -53,33 +54,29 @@
 	}
 
 	const init = async () => {
-		mode === 'cascade' &&
-			elements.forEach(({ currentNode }) => (currentNode.textContent = ''))
+		mode === 'cascade' && elements.forEach(({ currentNode }) => (currentNode.textContent = ''))
 		for (const element of elements) {
-			mode === 'cascade'
-				? await typewriterEffect(element)
-				: typewriterEffect(element)
+			mode === 'cascade' ? await typewriterEffect(element) : typewriterEffect(element)
 		}
 		dispatch('done')
 	}
 
 	const loop = async () => {
-		const loopParagraphTag = node.firstChild.tagName.toLowerCase()
-		const loopParagraph = document.createElement(loopParagraphTag)
-		node.childNodes.forEach(el => el.remove())
-		node.appendChild(loopParagraph)
 		while (mode === 'loop') {
-			for (const { text } of elements) {
+			for (const { currentNode, text } of elements) {
+				node.childNodes.forEach(el => el.remove())
+				const loopParagraphTag = currentNode.tagName.toLowerCase()
+				const loopParagraph = document.createElement(loopParagraphTag)
 				loopParagraph.textContent = text.join('')
+				node.appendChild(loopParagraph)
 				await typewriterEffect({ currentNode: loopParagraph, text })
+				node.childNodes.forEach(el => el.remove())
 			}
 			dispatch('done')
 		}
 	}
 
 	onMount(() => {
-		if (hasSingleTextNode(node))
-			throw new Error('<Typewriter /> must have at least one element')
 		getElements(node)
 		switch (mode) {
 			case 'cascade':
@@ -117,10 +114,6 @@
 	}
 </style>
 
-<div
-	class:cursor
-	style="--cursor-color: {typeof cursor === 'string' ? cursor : 'black'}"
-	bind:this={node}
->
+<div class:cursor style="--cursor-color: {typeof cursor === 'string' ? cursor : 'black'}" bind:this={node}>
 	<slot />
 </div>